feat: added database functionality and configurations
The commit includes adding new files for database configurations and utilities with PostgreSQL as the chosen database. It also includes establishing a new connection to PostgreSQL in the main.rs file. New modules for "utils" and "config" have also been added. The config module fetches the database URL from the environment variable. The database utility includes scripts for running migrations on the database.