feat(category-api): add categories api routes and handlers
Added new category-related routes and handlers in category/api module of the application. Handlers for creating, finding, deleting and updating a category are added. Created a config file to register these new routes. Creating a category requires the name of the category and the store_id. To find categories, store_id is used. Categories can be deleted and updated using category_id.