feat(product): incorporate store_id to product and price operations
In this commit, store_id is now passed into methods in the domain, db adapter and service layer for functionalities which operate on products and their images and prices. This feature is to ensure each operation belongs to the correct store and added a ‘Forbidden’ error handling for when the store_id does not matches. It includes:
- In the handlers.rs file, actix-session is imported to get user_id and passed to delete_product method to enhance security.
- In ports.rs, an additional argument of store_id is added to the deletion and updating, add_images, delete_image, and price handling methods in the Repository trait.
- In db_adapter.rs, added the validation of store_id which is fetched via actix-session.
- In service.rs, an additional argument store_id is added to the product and price handling methods.